Refactor internationalization: streamline locale handling, remove unused I18nProvider, and enhance Nav component
