+++ src/app/[locale]/layout.tsx
@@ -2,9 +2,8 @@ import type { Metadata } from 'next';
 import { notFound } from 'next/navigation';
 import './globals.css';
 
-import { hasLocale, locales, pickClientDictionary } from '@/i18n/config';
+import { hasLocale, locales } from '@/i18n/config';
 import { getDictionary } from '@/i18n/dictionaries';
-import { I18nProvider } from '@/i18n/client';
 import { Footer, Nav } from '@/components/layout';
 import { ToTop } from '@/components/layout';
 
@@ -23,17 +22,15 @@ export default async function RootLayout({
 }: LayoutProps<'/[locale]'>) {
   const { locale } = await params;
   if (!hasLocale(locale)) notFound();
-  const clientDictionary = pickClientDictionary(await getDictionary(locale));
+  const dict = await getDictionary(locale);
 
   return (
     <html lang={locale} suppressHydrationWarning>
       <body>
-        <I18nProvider locale={locale} dictionary={clientDictionary}>
-          <Nav />
-          <ToTop />
-          {children}
-          <Footer />
-        </I18nProvider>
+        <Nav labels={dict.navigation} />
+        <ToTop />
+        {children}
+        <Footer />
       </body>
     </html>
   );
